Reafactor WIP new subblock decoder

+++ spec/gd.h
@@ -26,4 +26,4 @@ typedef struct {
 void gd_open(gd_info_t *info);
 void gd_decode(gd_decode_t *decode);
 uint16_t gd_image_sub_block_decode(gd_image_data_block_decode_t *blockDecode, uint16_t *output);
-void gd_image_subblock_decode(gd_image_block_t *block, uint8_t *subblock, uint16_t count);
+void gd_image_subblock_decode(gd_image_block_t *block, uint8_t *subblock, uint8_t count);
